Enabled aggressive build and analyzer warnings.
* Turned on "treat warnings as errors"
* Promoted some build settings from target to project level
* Deleted redundant target-specific settings
* Turned on a raft of aggressive build warnings, fixed the few issues they turned up
* Fixed a handful of static analyzer issues, including some nullability warnings
* Added a few TODOs for nullability, where we're not conforming to a platform/3rd party protocol

+++ OBAKit/Helpers/OBAMapHelpers.m
@@ -286,7 +286,7 @@ NSInteger OBASortStopsByDistanceFromLocation(OBAStopV2 *stop1, OBAStopV2 *stop2,
 + (NSUInteger)zoomLevelForMapRect:(MKMapRect)mRect withMapViewSizeInPixels:(CGSize)viewSizeInPixels
 {
     NSUInteger zoomLevel = MAXIMUM_ZOOM;
-    MKZoomScale zoomScale = mRect.size.width / viewSizeInPixels.width; //MKZoomScale is just a CGFloat typedef
+    MKZoomScale zoomScale = (CGFloat) mRect.size.width / viewSizeInPixels.width; //MKZoomScale is just a CGFloat typedef
     double zoomExponent = log2(zoomScale);
     zoomLevel = (NSUInteger)(MAXIMUM_ZOOM - ceil(zoomExponent));
     return zoomLevel;
